+++ src/Domain/Common/Storage/LocalUploadStorage.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace PhpList\Core\Domain\Common\Upload\Storage;
+namespace PhpList\Core\Domain\Common\Storage;
 
-use PhpList\Core\Domain\Common\Upload\Exception\StorageException;
+use PhpList\Core\Domain\Common\Exception\StorageException;
+use PhpList\Core\Domain\Common\Model\StoredFile;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
