+++ src/Domain/Common/Storage/UploadStorageInterface.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace PhpList\Core\Domain\Common\Upload\Storage;
+namespace PhpList\Core\Domain\Common\Storage;
 
+use PhpList\Core\Domain\Common\Model\StoredFile;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 interface UploadStorageInterface
