+++ src/Domain/Common/Validator/UploadValidator.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace PhpList\Core\Domain\Common\Upload;
+namespace PhpList\Core\Domain\Common\Validator;
 
-use PhpList\Core\Domain\Common\Upload\Exception\InvalidUploadException;
-use PhpList\Core\Domain\Common\Upload\Exception\MissingUploadException;
-use PhpList\Core\Domain\Common\Upload\Exception\UnsupportedExtensionException;
-use PhpList\Core\Domain\Common\Upload\Exception\UnsupportedMimeTypeException;
-use PhpList\Core\Domain\Common\Upload\Exception\UploadTooLargeException;
+use PhpList\Core\Domain\Common\Exception\InvalidUploadException;
+use PhpList\Core\Domain\Common\Exception\MissingUploadException;
+use PhpList\Core\Domain\Common\Exception\UnsupportedExtensionException;
+use PhpList\Core\Domain\Common\Exception\UnsupportedMimeTypeException;
+use PhpList\Core\Domain\Common\Exception\UploadTooLargeException;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
