+++ src/Domain/Messaging/Service/TemplateImageEmbedder.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace PhpList\Core\Domain\Messaging\Service;
 
 use Exception;
-use PhpList\Core\Domain\Common\ExternalImageService;
 use PhpList\Core\Domain\Common\Model\ContentTransferEncoding;
+use PhpList\Core\Domain\Common\Service\ExternalImageService;
 use PhpList\Core\Domain\Configuration\Model\ConfigOption;
 use PhpList\Core\Domain\Configuration\Model\Dto\CreateConfigDto;
 use PhpList\Core\Domain\Configuration\Service\Manager\ConfigManager;
